Added move warnings
