+++ tda/debug.py
@@ -7,6 +7,7 @@ import tda
 def get_logger():
     return logging.getLogger(__name__)
 
+
 class LogRedactor:
     '''
     Collects strings that should not be emitted and replaces them with safe
@@ -95,9 +96,10 @@ def enable_bug_report_logging():
     '''
     _enable_bug_report_logging()
 
+
 def _enable_bug_report_logging(output=sys.stderr, loggers=None):
     '''
-    Module-internal version of :func:`enable_bug_report_logging`, intended for 
+    Module-internal version of :func:`enable_bug_report_logging`, intended for
     use in tests.
     '''
     if loggers is None:
