+++ tda/orders/equities.py
@@ -1,22 +1,6 @@
 from enum import Enum
 
-
-class InvalidOrderException(Exception):
-    '''Raised when attempting to build an incomplete order'''
-    pass
-
-
-class Duration(Enum):
-    DAY = 'DAY'
-    GOOD_TILL_CANCEL = 'GOOD_TILL_CANCEL'
-    FILL_OR_KILL = 'FILL_OR_KILL'
-
-
-class Session(Enum):
-    NORMAL = 'NORMAL'
-    AM = 'AM'
-    PM = 'PM'
-    SEAMESS = 'SEAMLESS'
+from tda.orders.common import InvalidOrderException, Duration, Session
 
 
 class EquityOrderBuilder:
